Lab1UVSS-feature-vizualizareNote/src/main/java/org/example/controller/CourseController.java
package org.example.controller;

import jakarta.persistence.EntityNotFoundException;
import org.example.model.Course;
import org.example.service.CourseService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/courses")
public class CourseController {
    @Autowired
    private CourseService courseService;

    @GetMapping
    public List<Course> getAllCourses() {
        return courseService.getAllCourses();
    }

    @GetMapping("/{id}")
    public ResponseEntity<Course> getCourseById(@PathVariable Integer id) {
        Course course = courseService.getCourseById(id).orElse(null);
        return (course != null) ? ResponseEntity.ok(course) : ResponseEntity.notFound().build();

    }


    @PostMapping
    public ResponseEntity<?> addCourse(@RequestBody Course course) {
        {
            Course addedCourse = courseService.addCourses(course);
            return ResponseEntity.status(HttpStatus.CREATED).body(addedCourse);

        }
    }


    @DeleteMapping("/{id}")
    public ResponseEntity<?> deleteCourse(@PathVariable Integer id) {
        try {
            courseService.deleteCourse(id);
            return ResponseEntity.noContent().build();
        } catch (EntityNotFoundException e) {
            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
        } catch (Exception e) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
        }
    }

}
